+++ src/translations/translations.fr.js
@@ -214,6 +214,28 @@ export const translations = {
   'cc-heptapod-info.not-in-use': `Vous n'utilisez pas ce service Heptapod.`,
   'cc-heptapod-info.description': () => sanitize`Cette instance Heptapod héberge des dépôts Mercurial. Plus d'informations sur <a href="https://about.heptapod.host" rel="noreferrer noopener">https://about.heptapod.host</a>.`,
   'cc-heptapod-info.error-loading': `Une erreur est survenue pendant le chargement des informations d'utilisation.`,
+  // cc-invoice
+  'cc-invoice.download-pdf': `Télécharger le PDF`,
+  'cc-invoice.error': `Une erreur est survenue pendant le chargement de la facture.`,
+  'cc-invoice.info': ({ date, amount }) => sanitize`Cette facture a été émise le <strong>${formatDateOnly(date)}</strong> pour un total de <strong>${currencyFormatter.format(amount)}</strong>.`,
+  'cc-invoice.title': `Facture`,
+  // cc-invoice-list
+  'cc-invoice-list.error': `Une erreur est survenue pendant le chargement des factures.`,
+  'cc-invoice-list.pending': `Factures en attente de paiement`,
+  'cc-invoice-list.pending.no-invoices': `Il n'y a aucune facture en attente de paiement pour le moment.`,
+  'cc-invoice-list.processed': `Factures réglées`,
+  'cc-invoice-list.processed.no-invoices': `Il n'y a aucune facture réglée pour le moment.`,
+  'cc-invoice-list.processing': `Factures dont le paiement est en cours de validation`,
+  'cc-invoice-list.title': `Factures`,
+  'cc-invoice-list.year': `Année :`,
+  // cc-invoice-table
+  'cc-invoice-table.date.payment': `Date de paiement`,
+  'cc-invoice-table.date.value': ({ date }) => `${formatDateOnly(date)}`,
+  'cc-invoice-table.number': `Numéro`,
+  'cc-invoice-table.open-pdf': `Télécharger le PDF`,
+  'cc-invoice-table.pay': `Régler`,
+  'cc-invoice-table.total.label': `Total`,
+  'cc-invoice-table.total.value': ({ amount }) => `${currencyFormatter.format(amount)}`,
   // cc-addon-admin
   'cc-addon-admin.addon-name': `Nom de l'add-on`,
   'cc-addon-admin.admin': `Administration`,