+++ CHANGELOG.md
@@ -6,7 +6,11 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
-...
+* New components:
+  * `<cc-html-frame>`
+  * `<cc-invoice>`
+  * `<cc-invoice-list>` (with smart definition)
+  * `<cc-invoice-table>`
 
 ## 6.1.0 (2021-03-25)